refactor(loyalty): keep counters in queries.go, flatten the transfer gate
The previous split moved the counter half into a new counters.go, which held
long-standing counter code to the health bar for new files and reported its
primitive-heavy signatures as fresh findings. The split is the same, the
direction is not: the ledger moves to balances.go and the counters stay where
they were.

pointsGive opened with a four-operand conditional that mixed the capability
gate with the amount parse, then re-tested the gate inside to pick a reply key.
The gate is now its own branch, and the parse plus bound it shared with
pointsAdjust becomes boundedAmount.
