+++ app/loyalty/repository/queries.go
@@ -5,7 +5,6 @@ package repository
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"strings"
 
@@ -13,303 +12,428 @@ import (
 	"ItsBagelBot/app/loyalty/ent/balance"
 	"ItsBagelBot/app/loyalty/ent/counter"
 	"ItsBagelBot/app/loyalty/ent/counterentry"
+	"ItsBagelBot/internal/domain/event/data"
 	"ItsBagelBot/pkg/db"
 
 	entsql "entgo.io/ent/dialect/sql"
 )
 
-// defaultTopLimit bounds a leaderboard read when the caller sent no (or a
-// silly) limit.
-const (
-	defaultTopLimit = 10
-	maxTopLimit     = 100
-)
+// maxCounterName mirrors the schema's MaxLen; enforced at the trust boundary
+// so a hostile RPC can never hit the column constraint as a DB error.
+const maxCounterName = 64
 
-// ErrInvalidInput marks trust-boundary rejections; the RPC layer maps it to a
-// "bad request" reply instead of a generic failure.
-var ErrInvalidInput = errors.New("invalid input")
+// ValidCounterName reports the normalized name, or an error when it is empty,
+// oversized, or contains ':' — reserved so the worker's "{counter:bot:name}"
+// token prefix can never collide with a stored counter name.
+func ValidCounterName(name string) (string, error) {
+	n := normalizeName(name)
+	if n == "" || len(n) > maxCounterName {
+		return "", fmt.Errorf("%w: counter name", ErrInvalidInput)
+	}
+	if strings.Contains(n, ":") {
+		return "", fmt.Errorf("%w: counter name", ErrInvalidInput)
+	}
+	return n, nil
+}
 
-// BalanceGet returns one viewer's standing. A missing row is (zero, false, nil).
-func (r *Loyalty) BalanceGet(ctx context.Context, userID, viewerID uint64) (*ent.Balance, bool, error) {
-	return getOptional(ctx, func(ctx context.Context) (*ent.Balance, error) {
+// writableCounterName validates a name for a write verb and, for a real
+// broadcaster, refuses the system-owned fleet stats names (data.SystemCounter).
+// Those rows are written for every channel by sesame and ranked on the public
+// stats boards, so letting a channel create, set, rename or delete its own row
+// would let it edit a public leaderboard. The bot namespace (user 0) keeps full
+// control: that is the admin console's counter surface.
+func writableCounterName(userID uint64, name string) (string, error) {
+	n, err := ValidCounterName(name)
+	if err != nil {
+		return "", err
+	}
+	if userID != 0 && data.SystemCounter(n) {
+		return "", fmt.Errorf("%w: reserved counter name", ErrInvalidInput)
+	}
+	return n, nil
+}
+
+// ValidScope reports the canonical scope, defaulting empty to channel.
+func ValidScope(scope string) (string, error) {
+	switch scope {
+	case "", data.CounterScopeChannel:
+		return data.CounterScopeChannel, nil
+	case data.CounterScopeBot, data.CounterScopeViewer, data.CounterScopeCommand, data.CounterScopeViewerCommand:
+		return scope, nil
+	default:
+		return "", fmt.Errorf("%w: scope", ErrInvalidInput)
+	}
+}
+
+// entryScoped reports whether a scope keeps its values in counter_entries;
+// bot and channel scopes keep the value on the counter row itself.
+func entryScoped(scope string) bool {
+	switch scope {
+	case data.CounterScopeViewer, data.CounterScopeCommand, data.CounterScopeViewerCommand:
+		return true
+	default:
+		return false
+	}
+}
+
+// bucketed reports whether a scope keys entries by command bucket.
+func bucketed(scope string) bool {
+	return scope == data.CounterScopeCommand || scope == data.CounterScopeViewerCommand
+}
+
+// CounterEntries lists an entry-scoped counter's stored buckets, highest
+// value first (the dashboard's per-counter leaderboard). Each bucket carries
+// its own stored viewer identity (refreshed by the bump flushes); the login
+// map fills the gap for legacy rows written before identity was stored, from
+// the viewer's balance row when they have one.
+func (r *Loyalty) CounterEntries(ctx context.Context, userID uint64, name string, limit int) ([]*ent.CounterEntry, map[uint64]string, error) {
+	n, err := ValidCounterName(name)
+	if err != nil {
+		return nil, nil, err
+	}
+	rows, err := db.WithQuery(ctx, func(ctx context.Context) ([]*ent.CounterEntry, error) {
+		return r.client.CounterEntry.Query().
+			Where(counterentry.UserIDEQ(userID), counterentry.NameEQ(n)).
+			Order(counterentry.ByValue(entsql.OrderDesc()), counterentry.ByViewerID()).
+			Limit(clampLimit(limit)).
+			All(ctx)
+	})
+	if err != nil || len(rows) == 0 {
+		return rows, nil, err
+	}
+	return rows, r.viewerLogins(ctx, userID, rows), nil
+}
+
+// viewerLogins resolves the display login of each distinct viewer whose row
+// carries no stored identity yet, from their balance row. Best-effort: logins
+// are cosmetic, the entries themselves are the answer, so a read failure
+// returns an empty map.
+func (r *Loyalty) viewerLogins(ctx context.Context, userID uint64, rows []*ent.CounterEntry) map[uint64]string {
+	logins := map[uint64]string{}
+	ids := legacyViewerIDs(rows)
+	if len(ids) == 0 {
+		return logins
+	}
+	bals, err := db.WithQuery(ctx, func(ctx context.Context) ([]*ent.Balance, error) {
 		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID), balance.ViewerIDEQ(viewerID)).
+			Where(balance.UserIDEQ(userID), balance.ViewerIDIn(ids...)).
+			All(ctx)
+	})
+	if err != nil {
+		return logins
+	}
+	for _, b := range bals {
+		if b.ViewerLogin != "" {
+			logins[b.ViewerID] = b.ViewerLogin
+		}
+	}
+	return logins
+}
+
+// legacyViewerIDs returns the distinct viewers in rows whose bucket carries no
+// stored login yet — the only ones that still need a balance-row lookup.
+func legacyViewerIDs(rows []*ent.CounterEntry) []uint64 {
+	seen := map[uint64]struct{}{}
+	ids := make([]uint64, 0, len(rows))
+	for _, e := range rows {
+		if e.ViewerID == 0 || e.ViewerLogin != "" {
+			continue
+		}
+		if _, dup := seen[e.ViewerID]; !dup {
+			seen[e.ViewerID] = struct{}{}
+			ids = append(ids, e.ViewerID)
+		}
+	}
+	return ids
+}
+
+// entryTarget resolves how a read or write addresses one of row's entry
+// buckets: the (viewer, command) pair when the caller targeted one, ok=false
+// when untargeted. Untargeted means the row value answers a read and a set
+// resets the whole counter. Command scope targets by command alone (pooled
+// across viewers); the viewer scopes target by viewer; channel/bot never
+// target entries.
+func entryTarget(row *ent.Counter, viewerID uint64, command string) (uint64, string, bool) {
+	switch row.Scope {
+	case data.CounterScopeCommand:
+		cmd := normalizeCommand(command)
+		return 0, cmd, cmd != ""
+	case data.CounterScopeViewer:
+		return viewerID, "", viewerID != 0
+	case data.CounterScopeViewerCommand:
+		return viewerID, normalizeCommand(command), viewerID != 0
+	default:
+		return 0, "", false
+	}
+}
+
+// CounterGet resolves one counter: the definition, plus the effective value —
+// the row's own value for channel/bot scope; the selected bucket's value (0
+// when it has none) for the entry scopes: the viewer's entry for viewer
+// scope, the command bucket (pooled, viewer-independent) for command scope,
+// the (command, viewer) bucket for viewer+command scope. A viewer scope asked
+// without a viewer answers with the row value.
+func (r *Loyalty) CounterGet(ctx context.Context, userID uint64, name string, viewerID uint64, command string) (*ent.Counter, int64, bool, error) {
+	n, err := ValidCounterName(name)
+	if err != nil {
+		return nil, 0, false, err
+	}
+	row, found, err := getOptional(ctx, func(ctx context.Context) (*ent.Counter, error) {
+		return r.client.Counter.Query().
+			Where(counter.UserIDEQ(userID), counter.NameEQ(n)).
 			Only(ctx)
+	})
+	if err != nil || !found {
+		return nil, 0, found, err
+	}
+	entryViewer, cmd, useEntry := entryTarget(row, viewerID, command)
+	if !useEntry {
+		return row, row.Value, true, nil
+	}
+	entry, entryFound, err := getOptional(ctx, func(ctx context.Context) (*ent.CounterEntry, error) {
+		return r.client.CounterEntry.Query().
+			Where(
+				counterentry.UserIDEQ(userID),
+				counterentry.NameEQ(n),
+				counterentry.CommandEQ(cmd),
+				counterentry.ViewerIDEQ(entryViewer),
+			).
+			Only(ctx)
+	})
+	if err != nil {
+		return nil, 0, false, err
+	}
+	if !entryFound {
+		return row, 0, true, nil
+	}
+	return row, entry.Value, true, nil
+}
+
+// CountersList returns the channel's counter definitions. The system-owned
+// fleet stats rows sesame writes on every channel are left out: the broadcaster
+// cannot edit them (writableCounterName), so listing them would only offer a
+// dashboard row whose every action fails. The bot namespace still sees them —
+// that is where the admin console manages the fleet totals.
+func (r *Loyalty) CountersList(ctx context.Context, userID uint64) ([]*ent.Counter, error) {
+	return db.WithQuery(ctx, func(ctx context.Context) ([]*ent.Counter, error) {
+		q := r.client.Counter.Query().Where(counter.UserIDEQ(userID))
+		if userID != 0 {
+			q = q.Where(counter.NameNotIn(data.CounterMessagesProcessed, data.CounterEventsProcessed))
+		}
+		return q.Order(counter.ByName()).All(ctx)
 	})
 }
 
-// Top returns the channel's top standings by points.
-func (r *Loyalty) Top(ctx context.Context, userID uint64, limit int) ([]*ent.Balance, error) {
-	return db.WithQuery(ctx, func(ctx context.Context) ([]*ent.Balance, error) {
-		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID)).
-			Order(balance.ByPoints(entsql.OrderDesc()), balance.ByViewerID()).
+// CounterBoard ranks every channel that holds the named channel-scope counter,
+// highest value first — the cross-broadcaster read behind the public stats
+// boards, where one counter name (say "messages_processed") is written by every
+// channel and the interesting answer is who leads.
+//
+// The reserved bot namespace (user 0) is excluded: it holds the fleet-wide
+// total of the same counter name, which would otherwise take rank 1 forever.
+// Entry-scoped rows are excluded too — their Value column is an unused 0.
+//
+// Backed by the (name, value) index; without it this is a full scan of the
+// counters table, which is why the index ships with this query.
+func (r *Loyalty) CounterBoard(ctx context.Context, name string, limit int) ([]*ent.Counter, error) {
+	n, err := ValidCounterName(name)
+	if err != nil {
+		return nil, err
+	}
+	return db.WithQuery(ctx, func(ctx context.Context) ([]*ent.Counter, error) {
+		return r.client.Counter.Query().
+			Where(
+				counter.NameEQ(n),
+				counter.UserIDNEQ(0),
+				counter.ScopeEQ(data.CounterScopeChannel),
+			).
+			Order(counter.ByValue(entsql.OrderDesc()), counter.ByUserID()).
 			Limit(clampLimit(limit)).
 			All(ctx)
 	})
 }
 
-// BalanceAdjust writes a viewer's points by login (a mod's "!points set/add
-// @user" — chat knows the target's login, not their id). absolute sets the
-// value; otherwise value is a delta. The row must already exist (any accrual
-// creates it); an unseen login is (nil, false, nil) so the caller can answer
-// "haven't seen them yet" instead of inventing an id-less row. Renames can
-// leave several rows carrying one old login; the freshest wins.
-func (r *Loyalty) BalanceAdjust(ctx context.Context, userID uint64, viewerLogin string, value int64, absolute bool) (*ent.Balance, bool, error) {
-	login := strings.ToLower(strings.TrimPrefix(strings.TrimSpace(viewerLogin), "@"))
-	if login == "" {
-		return nil, false, fmt.Errorf("%w: viewer_login", ErrInvalidInput)
+// CounterCreate upserts a counter definition. An existing counter keeps its
+// value and scope (create is idempotent, not a reset).
+func (r *Loyalty) CounterCreate(ctx context.Context, userID uint64, name, scope string) (*ent.Counter, error) {
+	n, err := writableCounterName(userID, name)
+	if err != nil {
+		return nil, err
 	}
-	row, found, err := getOptional(ctx, func(ctx context.Context) (*ent.Balance, error) {
-		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID), balance.ViewerLoginEQ(login)).
-			Order(balance.ByUpdatedAt(entsql.OrderDesc()), balance.ByViewerID()).
-			First(ctx)
-	})
-	if err != nil || !found {
-		return nil, found, err
+	s, err := ValidScope(scope)
+	if err != nil {
+		return nil, err
 	}
-	return row, true, db.WithExec(ctx, func(ctx context.Context) error {
-		upd := r.client.Balance.UpdateOneID(row.ID)
-		if absolute {
-			row.Points = value
-			upd.SetPoints(value)
-		} else {
-			row.Points += value
-			upd.AddPoints(value)
+	// Bot scope and the bot namespace imply each other: user 0 holds only
+	// bot counters, and bot counters live only under user 0.
+	if (s == data.CounterScopeBot) != (userID == 0) {
+		return nil, fmt.Errorf("%w: scope", ErrInvalidInput)
+	}
+	return db.WithQuery(ctx, func(ctx context.Context) (*ent.Counter, error) {
+		err := r.client.Counter.Create().
+			SetUserID(userID).
+			SetName(n).
+			SetScope(s).
+			OnConflict(entsql.ConflictColumns(counter.FieldUserID, counter.FieldName)).
+			Ignore().
+			Exec(ctx)
+		if err != nil {
+			return nil, err
 		}
-		return upd.Exec(ctx)
-	})
-}
-
-// BalanceSpend conditionally debits amount points from a viewer, addressed by
-// login (chat wagers know the target's login, not their id). The guard lives
-// in the UPDATE's WHERE clause — points >= amount — not in a prior read, so a
-// concurrent second spend racing the same row can never drive points
-// negative: whichever UPDATE lands second no longer matches and is refused.
-// spent=false with found=false means the channel never accrued for that
-// login; spent=false with found=true means insufficient points (Balance then
-// carries what they actually hold). amount must be positive: a zero or
-// negative "spend" is a caller bug, not a refund path.
-func (r *Loyalty) BalanceSpend(ctx context.Context, userID uint64, viewerLogin string, amount int64) (*ent.Balance, bool, bool, error) {
-	login, err := normalizeSpendTarget(viewerLogin, amount)
-	if err != nil {
-		return nil, false, false, err
-	}
-	row, found, err := getOptional(ctx, func(ctx context.Context) (*ent.Balance, error) {
-		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID), balance.ViewerLoginEQ(login)).
-			Order(balance.ByUpdatedAt(entsql.OrderDesc()), balance.ByViewerID()).
-			First(ctx)
-	})
-	if err != nil || !found {
-		return nil, false, found, err
-	}
-	n, err := r.client.Balance.Update().
-		Where(balance.IDEQ(row.ID), balance.PointsGTE(amount)).
-		AddPoints(-amount).
-		Save(ctx)
-	if err != nil {
-		return nil, true, false, err
-	}
-	if n == 0 {
-		// Refused: another write moved the balance after our read, so the
-		// snapshot above is stale. Report what the row holds now.
-		fresh, ferr := r.client.Balance.Get(ctx, row.ID)
-		if ferr != nil {
-			return nil, true, false, ferr
-		}
-		return fresh, true, false, nil
-	}
-	spent, err := r.client.Balance.Get(ctx, row.ID)
-	if err != nil {
-		return nil, true, true, err
-	}
-	return spent, true, true, nil
-}
-
-// normalizeSpendTarget validates a spend request and returns the lower-cased
-// login the ledger addresses.
-func normalizeSpendTarget(viewerLogin string, amount int64) (string, error) {
-	login := strings.ToLower(strings.TrimPrefix(strings.TrimSpace(viewerLogin), "@"))
-	if login == "" || amount <= 0 {
-		return "", fmt.Errorf("%w: viewer_login/amount", ErrInvalidInput)
-	}
-	return login, nil
-}
-
-// errInsufficient marks the guarded debit that matched no row — an internal
-// sentinel unwound out of the transaction, never surfaced as itself.
-var errInsufficient = errors.New("insufficient points")
-
-// TransferOutcome is one balance.transfer's result: the sender's and
-// recipient's standings after the move (To is non-nil only when moved).
-type TransferOutcome struct {
-	From, To *ent.Balance
-}
-
-// BalanceTransfer moves amount points from one viewer to another in one
-// transaction: the debit is guarded by points >= amount (same shape as
-// BalanceSpend, so a concurrent second spend can never drive the sender
-// negative), and the credit lands on the recipient row resolved by login —
-// which must already exist, exactly like a mod grant. found=false means the
-// channel never accrued for sender or target; moved=false with found=true
-// means the sender could not cover it (From then carries their actual
-// standing). A self-transfer is refused up front: the net-zero move has no
-// caller story, so it is input, not ledger state.
-func (r *Loyalty) BalanceTransfer(ctx context.Context, userID, fromViewerID uint64, targetLogin string, amount int64) (*TransferOutcome, bool, error) {
-	sender, recipient, found, err := r.transferParties(ctx, userID, fromViewerID, targetLogin, amount)
-	if err != nil || !found {
-		return nil, found, err
-	}
-
-	err = r.moveBalance(ctx, sender.ID, recipient.ID, amount)
-	if errors.Is(err, errInsufficient) {
-		// The guarded debit matched no row: report the sender's real standing
-		// rather than the pre-flight read, which a concurrent spend may have
-		// already invalidated.
-		fresh, ferr := r.client.Balance.Get(ctx, sender.ID)
-		if ferr != nil {
-			return nil, true, ferr
-		}
-		return &TransferOutcome{From: fresh}, true, nil
-	}
-	if err != nil {
-		return nil, true, err
-	}
-	return r.transferOutcome(ctx, sender.ID, recipient.ID)
-}
-
-// transferParties resolves and validates both sides of a move. found=false
-// means one of them never accrued in this channel; a self-transfer is input,
-// not ledger state, so it errors rather than performing a net-zero write.
-func (r *Loyalty) transferParties(ctx context.Context, userID, fromViewerID uint64, targetLogin string, amount int64) (sender, recipient *ent.Balance, found bool, err error) {
-	login, err := normalizeSpendTarget(targetLogin, amount)
-	if err != nil {
-		return nil, nil, false, err
-	}
-	if fromViewerID == 0 {
-		return nil, nil, false, fmt.Errorf("%w: from_viewer_id", ErrInvalidInput)
-	}
-
-	sender, found, err = getOptional(ctx, func(ctx context.Context) (*ent.Balance, error) {
-		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID), balance.ViewerIDEQ(fromViewerID)).
+		return r.client.Counter.Query().
+			Where(counter.UserIDEQ(userID), counter.NameEQ(n)).
 			Only(ctx)
 	})
-	if err != nil || !found {
-		return nil, nil, found, err
-	}
-	recipient, found, err = getOptional(ctx, func(ctx context.Context) (*ent.Balance, error) {
-		return r.client.Balance.Query().
-			Where(balance.UserIDEQ(userID), balance.ViewerLoginEQ(login)).
-			Order(balance.ByUpdatedAt(entsql.OrderDesc()), balance.ByViewerID()).
-			First(ctx)
-	})
-	if err != nil || !found {
-		return nil, nil, found, err
-	}
-	if recipient.ID == sender.ID {
-		return nil, nil, true, fmt.Errorf("%w: self transfer", ErrInvalidInput)
-	}
-	return sender, recipient, true, nil
 }
 
-// moveBalance is the debit and credit as one transaction. The debit carries
-// the points >= amount predicate (the same shape BalanceSpend uses) so a
-// concurrent second spend can never drive the sender negative; matching no row
-// unwinds the transaction with errInsufficient, leaving the credit unapplied.
-func (r *Loyalty) moveBalance(ctx context.Context, senderID, recipientID int, amount int64) error {
-	return db.WithExec(ctx, func(ctx context.Context) error {
+// CounterSet writes an absolute value. Channel/bot scope sets the row value.
+// For the entry scopes, a targeted set upserts that bucket; an untargeted set
+// resets the whole counter (deletes every entry), the "!counter reset"
+// semantics. A missing counter is (false, nil).
+func (r *Loyalty) CounterSet(ctx context.Context, userID uint64, name string, target SetTarget, value int64) (bool, error) {
+	if _, err := writableCounterName(userID, name); err != nil {
+		return false, err
+	}
+	row, _, found, err := r.CounterGet(ctx, userID, name, 0, "")
+	if err != nil || !found {
+		return found, err
+	}
+	return true, db.WithExec(ctx, func(ctx context.Context) error {
+		if !entryScoped(row.Scope) {
+			return r.client.Counter.Update().
+				Where(counter.UserIDEQ(userID), counter.NameEQ(row.Name)).
+				SetValue(value).
+				Exec(ctx)
+		}
+		entryViewer, cmd, targeted := entryTarget(row, target.ViewerID, target.Command)
+		if !targeted {
+			_, err := r.client.CounterEntry.Delete().
+				Where(counterentry.UserIDEQ(userID), counterentry.NameEQ(row.Name)).
+				Exec(ctx)
+			return err
+		}
+		login := normalizeLogin(target.ViewerLogin)
+		return r.client.CounterEntry.Create().
+			SetUserID(userID).
+			SetName(row.Name).
+			SetCommand(cmd).
+			SetViewerID(entryViewer).
+			SetViewerLogin(login).
+			SetValue(value).
+			OnConflict(entsql.ConflictColumns(counterentry.FieldUserID, counterentry.FieldName, counterentry.FieldCommand, counterentry.FieldViewerID)).
+			Update(func(u *ent.CounterEntryUpsert) {
+				u.UpdateValue()
+				if login != "" {
+					u.UpdateViewerLogin()
+				}
+			}).
+			Exec(ctx)
+	})
+}
+
+// CounterEntryDelete removes one stored bucket of an entry-scoped counter,
+// addressed like a targeted set (a viewer for the viewer scopes, a command
+// bucket for command scope). It refuses an untargeted call so it can never
+// become an accidental whole-counter reset — that is CounterSet's job. A
+// missing counter, a non-entry scope, or an untargeted address is
+// (false, nil); an already-absent bucket is (true, nil), since the goal state
+// (no such bucket) holds either way. The worker's live Valkey view converges
+// the same way a delete does: TTL expiry, or re-seed on the next cold read.
+func (r *Loyalty) CounterEntryDelete(ctx context.Context, userID uint64, name string, target SetTarget) (bool, error) {
+	if _, err := writableCounterName(userID, name); err != nil {
+		return false, err
+	}
+	row, _, found, err := r.CounterGet(ctx, userID, name, 0, "")
+	if err != nil || !found {
+		return found, err
+	}
+	if !entryScoped(row.Scope) {
+		return false, nil
+	}
+	entryViewer, cmd, targeted := entryTarget(row, target.ViewerID, target.Command)
+	if !targeted {
+		return false, nil
+	}
+	return true, db.WithExec(ctx, func(ctx context.Context) error {
+		_, err := r.client.CounterEntry.Delete().
+			Where(
+				counterentry.UserIDEQ(userID),
+				counterentry.NameEQ(row.Name),
+				counterentry.CommandEQ(cmd),
+				counterentry.ViewerIDEQ(entryViewer),
+			).
+			Exec(ctx)
+		return err
+	})
+}
+
+// normalizeLogin canonicalizes a carried viewer login the way BalanceAdjust
+// does its target: bare, lower-cased, clamped to the column width.
+func normalizeLogin(login string) string {
+	l := strings.ToLower(strings.TrimPrefix(strings.TrimSpace(login), "@"))
+	if len(l) > maxCounterName {
+		l = l[:maxCounterName]
+	}
+	return l
+}
+
+// CounterRename moves a counter and its entry buckets to a new name, in one
+// transaction so a crash can never strand the buckets under the old name. The
+// target name must be free: a clash returns ErrInvalidInput so the caller can
+// surface "name taken" instead of a generic failure. A missing counter is
+// (false, nil). The worker's live Valkey view of the old name converges the
+// same way a delete does: TTL expiry, or re-seed from the service on the next
+// cold read.
+func (r *Loyalty) CounterRename(ctx context.Context, userID uint64, name, newName string) (bool, error) {
+	n, err := writableCounterName(userID, name)
+	if err != nil {
+		return false, err
+	}
+	nn, err := writableCounterName(userID, newName)
+	if err != nil || nn == n {
+		return false, fmt.Errorf("%w: new name", ErrInvalidInput)
+	}
+	renamed := false
+	err = db.WithExec(ctx, func(ctx context.Context) error {
 		return withTx(ctx, r.client, func(tx *ent.Tx) error {
-			updated, err := tx.Balance.Update().
-				Where(balance.IDEQ(senderID), balance.PointsGTE(amount)).
-				AddPoints(-amount).
+			updated, err := tx.Counter.Update().
+				Where(counter.UserIDEQ(userID), counter.NameEQ(n)).
+				SetName(nn).
 				Save(ctx)
 			if err != nil {
+				if ent.IsConstraintError(err) {
+					return fmt.Errorf("%w: name taken", ErrInvalidInput)
+				}
 				return err
 			}
 			if updated == 0 {
-				return errInsufficient
+				return nil
 			}
-			_, err = tx.Balance.Update().
-				Where(balance.IDEQ(recipientID)).
-				AddPoints(amount).
-				Save(ctx)
-			return err
+			renamed = true
+			return tx.CounterEntry.Update().
+				Where(counterentry.UserIDEQ(userID), counterentry.NameEQ(n)).
+				SetName(nn).
+				Exec(ctx)
 		})
 	})
+	return renamed, err
 }
 
-// transferOutcome re-reads both rows after a committed move, so the reply
-// carries post-move standings rather than the caller's arithmetic.
-func (r *Loyalty) transferOutcome(ctx context.Context, senderID, recipientID int) (*TransferOutcome, bool, error) {
-	from, err := r.client.Balance.Get(ctx, senderID)
-	if err != nil {
-		return nil, true, err
-	}
-	to, err := r.client.Balance.Get(ctx, recipientID)
-	if err != nil {
-		return nil, true, err
-	}
-	return &TransferOutcome{From: from, To: to}, true, nil
-}
-
-// clampLimit bounds a caller-provided page size, defaulting a missing one.
-func clampLimit(limit int) int {
-	if limit <= 0 {
-		return defaultTopLimit
-	}
-	return min(limit, maxTopLimit)
-}
-
-// SetTarget addresses one stored bucket of an entry-scoped counter (a viewer
-// for the viewer scopes, a command bucket for command scope). ViewerLogin
-// optionally stamps the bucket's display identity — the dashboard's manual
-// add knows the typed username; a later bump refreshes it like any other.
-type SetTarget struct {
-	ViewerID    uint64
-	Command     string
-	ViewerLogin string
-}
-
-// withTx runs fn inside one ent transaction, committing on success and
-// rolling back on error.
-func withTx(ctx context.Context, client *ent.Client, fn func(tx *ent.Tx) error) error {
-	tx, err := client.Tx(ctx)
+// CounterDelete removes a counter and its viewer entries.
+func (r *Loyalty) CounterDelete(ctx context.Context, userID uint64, name string) error {
+	n, err := writableCounterName(userID, name)
 	if err != nil {
 		return err
 	}
-	if err := fn(tx); err != nil {
-		_ = tx.Rollback()
-		return err
-	}
-	return tx.Commit()
-}
-
-// DeleteAllForUser removes every loyalty row of a deleted broadcaster account.
-func (r *Loyalty) DeleteAllForUser(ctx context.Context, userID uint64) error {
 	return db.WithExec(ctx, func(ctx context.Context) error {
-		if _, err := r.client.Balance.Delete().Where(balance.UserIDEQ(userID)).Exec(ctx); err != nil {
+		if _, err := r.client.CounterEntry.Delete().
+			Where(counterentry.UserIDEQ(userID), counterentry.NameEQ(n)).
+			Exec(ctx); err != nil {
 			return err
 		}
-		if _, err := r.client.CounterEntry.Delete().Where(counterentry.UserIDEQ(userID)).Exec(ctx); err != nil {
-			return err
-		}
-		_, err := r.client.Counter.Delete().Where(counter.UserIDEQ(userID)).Exec(ctx)
+		_, err := r.client.Counter.Delete().
+			Where(counter.UserIDEQ(userID), counter.NameEQ(n)).
+			Exec(ctx)
 		return err
 	})
-}
-
-// getOptional runs one Only-style query through the DB gate and maps ent's
-// not-found to (zero, false, nil).
-func getOptional[T any](ctx context.Context, fn func(context.Context) (*T, error)) (*T, bool, error) {
-	row, err := db.WithQuery(ctx, fn)
-	if err != nil {
-		if ent.IsNotFound(err) {
-			return nil, false, nil
-		}
-		return nil, false, err
-	}
-	return row, true, nil
 }
